fix: dropColumn('company') in migration

+++ database/migrations/2021_09_10_174152_alter_customer_table_company_field.php
@@ -27,7 +27,10 @@ class AlterCustomerTableCompanyField extends Migration
     public function down()
     {
         Schema::table('customer', function (Blueprint $table) {
-            $table->dropColumn('company');
+            if (Schema::hasColumn('customer', 'company')) {
+                $table->dropColumn('company');
+            }
+            
             $table->dropColumn('website');
         });
     }
